pgespresso_start_backup: fix crash on 9.5 databases with tablespaces
The do_pg_start_backup() function in PG 9.5 requires a pointer to a table
space map file, similar to the backup label.  The pointer isn't used in case
the database defines no tablespaces, but with tablespaces the pointer is
accessed unconditionally causing pgespresso to crash the server.

As explained in xlog.c's do_pg_start_backup comments the tablespace map file
isn't particularly useful on non-windows systems and as there's no easy way
to return it without changing the pgespresso_start_backup function API we
just ignore it.  The caller should be able to resolve tablespaces and handle
them cleanly.